refactor: 부모에서 props로 계산 결과 전달
