+++ src/components/ShoppingCartSection/ShoppingCartSection.tsx
@@ -1,5 +1,4 @@
-import { Dispatch, SetStateAction, useEffect } from "react";
-import { CartItemsResponse, Content } from "../../types/cartItems";
+import { CartItemsResponse } from "../../types/cartItems";
 import CartItem from "../CartItem/CartItem";
 import Checkbox from "../Checkbox/Checkbox";
 import InfoIcon from "../icons/Info";
@@ -13,6 +12,9 @@ interface ShoppingCartSectionProps {
   selectedItemIds: number[];
   onSelectItem: (itemId: number) => void;
   onSelectAll: () => void;
+  orderPrice: number;
+  shippingFee: number;
+  totalPrice: number;
 }
 
 export default function ShoppingCartSection({
@@ -21,25 +23,13 @@ export default function ShoppingCartSection({
   selectedItemIds,
   onSelectItem,
   onSelectAll,
+  orderPrice,
+  shippingFee,
+  totalPrice,
 }: ShoppingCartSectionProps) {
   const isAllSelected =
     shopppingCartItems?.content.length > 0 &&
     selectedItemIds.length === shopppingCartItems.content.length;
-
-  const orderPrice =
-    shopppingCartItems?.content.reduce((total, item) => {
-      const calculateItemPrice = (item: Content) => {
-        if (selectedItemIds.includes(item.id)) {
-          return item.product.price * item.quantity;
-        }
-        return 0;
-      };
-
-      return total + calculateItemPrice(item);
-    }, 0) || 0;
-  const shippingFee = orderPrice >= 100000 ? 0 : 3000;
-  const totalPrice = orderPrice + shippingFee;
-
 
   return (
     <S.ShoppingCartSection>
